requirements.txt
asgiref==3.9.1
click==8.2.1
Django==4.2.7
django-cors-headers==4.3.1
django-jazzmin==3.0.1
djangorestframework==3.14.0
dlib==20.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
numpy==2.0.0
opencv-contrib-python==4.12.0.88
opencv-python==4.8.1.78
Pillow==10.1.0
pytz==2025.2
setuptools==80.9.0
sqlparse==0.5.3
wheel==0.45.1
